+++ src/main/java/org/oa/mindbook/Dto/response/Memoir/JoyMemoirListResponseDto.java
@@ -1,4 +1,4 @@
-package org.oa.mindbook.Dto.response;
+package org.oa.mindbook.Dto.response.Memoir;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,37 +6,24 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.oa.mindbook.Domain.Entity.Memoir.JoyMemoir;
 
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
 @Builder
-public class JoyMemoirResponseDto {
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class JoyMemoirListResponseDto {
 
     private Long joyMemoirId;
 
-    private String userId;
-
     private String createdAt;
-
-    private String memory;
-
-    private String impression;
 
     private String status;
 
-    public static JoyMemoirResponseDto of(JoyMemoir joyMemoir) {
-        return JoyMemoirResponseDto.builder()
+    public static JoyMemoirListResponseDto of(JoyMemoir joyMemoir) {
+        return JoyMemoirListResponseDto.builder()
                 .joyMemoirId(joyMemoir.getJoyMemoirId())
-                .userId(joyMemoir.getUserId())
                 .createdAt(joyMemoir.getCreatedAt())
-                .memory(joyMemoir.getMemory())
-                .impression(joyMemoir.getImpression())
                 .status(joyMemoir.getStatus())
                 .build();
     }
-
-
-
-
 
 }
