+++ src/main/java/org/oa/mindbook/Repository/Memoir/JoyMemoirRepository.java
@@ -4,6 +4,9 @@ import org.oa.mindbook.Domain.Entity.Memoir.JoyMemoir;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface JoyMemoirRepository extends JpaRepository<JoyMemoir, Long> {
+    List<JoyMemoir> findByStatus(String status);
 }
