+++ src/main/java/org/oa/mindbook/Service/Memoir/JoyMemoirService.java
@@ -4,10 +4,15 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.oa.mindbook.Domain.Entity.Memoir.JoyMemoir;
-import org.oa.mindbook.Dto.request.CreateJoyMemoirRequestDto;
-import org.oa.mindbook.Dto.response.JoyMemoirResponseDto;
+import org.oa.mindbook.Dto.request.Memoir.CreateJoyMemoirRequestDto;
+import org.oa.mindbook.Dto.response.Memoir.JoyMemoirListResponseDto;
+import org.oa.mindbook.Dto.response.Memoir.JoyMemoirResponseDto;
 import org.oa.mindbook.Repository.Memoir.JoyMemoirRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.Dictionary;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -30,5 +35,16 @@ public class JoyMemoirService {
         JoyMemoir joyMemoir = joyMemoirRepository.findById(joyMemoirId).orElseThrow();
 
         return JoyMemoirResponseDto.of(joyMemoir);
+    }
+
+    @Transactional
+    public List<JoyMemoirListResponseDto> getJoyMemoirList(String status) {
+        List<JoyMemoir> joyMemoir = joyMemoirRepository.findByStatus(status);
+
+        List<JoyMemoirListResponseDto> responseDtoList = joyMemoir.stream()
+                .map(JoyMemoirListResponseDto::of)
+                .collect(Collectors.toList());
+
+        return responseDtoList;
     }
 }
